@@ -1,14 +1,16 @@
 plugins {
-	id 'org.springframework.boot' version '2.7.5'
-	id 'io.spring.dependency-management' version '1.0.15.RELEASE'
+	id 'org.springframework.boot' version '3.0.2'
+	id 'io.spring.dependency-management' version '1.1.0'
 	id 'java'
 }
 
 group = 'com.vmware.tanzu.apps.sso'
 version = '0.0.1-SNAPSHOT'
-sourceCompatibility = '11'
+sourceCompatibility = '17'
 
 ext['snakeyaml.version'] = '1.32'
+ext["thymeleaf.version"] = "3.1.1.RELEASE"
+ext["thymeleaf-extras-springsecurity.version"] = "3.1.1.RELEASE"
 
 repositories {
 	mavenCentral()
@@ -19,7 +21,7 @@ dependencies {
 	implementation 'org.springframework.boot:spring-boot-starter-thymeleaf'
 
 	implementation 'org.springframework.boot:spring-boot-starter-oauth2-client'
-	implementation 'org.thymeleaf.extras:thymeleaf-extras-springsecurity5'
+	implementation 'org.thymeleaf.extras:thymeleaf-extras-springsecurity6'
 
 	implementation "org.webjars:webjars-locator-core:0.50"
 	implementation "org.webjars:bootstrap:5.1.3"
