@@ -1,7 +1,7 @@
 package com.vmware.tanzu.apps.sso.accelerator.config;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.oauth2.client.oidc.web.logout.OidcClientInitiatedLogoutSuccessHandler;
@@ -12,10 +12,14 @@ import org.springframework.security.web.authentication.logout.LogoutSuccessHandl
 
 import static org.springframework.security.config.Customizer.withDefaults;
 
+@Configuration
 @EnableWebSecurity
 public class WebSecurityConfig {
-	@Autowired
 	private ClientRegistrationRepository clientRegistrationRepository;
+
+	public WebSecurityConfig(ClientRegistrationRepository clientRegistrationRepository) {
+		this.clientRegistrationRepository = clientRegistrationRepository;
+	}
 
 	@Bean
 	SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
@@ -26,7 +30,7 @@ public class WebSecurityConfig {
 				.authorizeRequests(authorizeRequests ->
 						authorizeRequests
 								// Permit all public paths explicitly.
-								.antMatchers("/", "/home", "/webjars/**", "/styles/**", "/images/**").permitAll()
+								.requestMatchers("/", "/home", "/webjars/**", "/styles/**", "/images/**").permitAll()
 								// Require authentication for all others paths.
 								.anyRequest().authenticated()
 				)
